Update new_number type to long int for palindrome check
Change new_number type from int to long int to handle larger values.

+++ palindrome_number.cpp
@@ -33,7 +33,7 @@ public:
         }
         
         int original = x;
-        int new_number = 0;
+        long int new_number = 0;
 
         while (x > 0) {
             int digit = x % 10;
